Updated logfile locations
Updated program so that the log file is created in the same folder as the script. This requires an absolute path since the intended use is with a cron job. Withoout an absolute path, the log file gets created in the home folder of the cronjob owner. I wanted more consistency.